destroy only if you are user, and rspec

+++ app/controllers/api/v1/projects_controller.rb
@@ -48,8 +48,11 @@ class Api::V1::ProjectsController < ApplicationController
 
   def destroy
     project = Project.find(params[:id])
-    project.destroy
-    render :nothing => true, :status => 200 #ok
+    if project.owner == self.getCurrentUser.id 
+      project.destroy
+      render :nothing => true, :status => :ok #200
+    else 
+      render :nothing => true, :status => :unauthorized
   end
 
   def getCurrentUser
